Fix missing comma in .clang-tidy and resolve newly-enabled bugprone warnings
The .clang-tidy Checks field uses a YAML folded scalar (>), which joins
lines with spaces. A missing comma after `bugprone-*` caused it to
concatenate with the next line into the invalid token `bugprone-* misc-*`,
silently disabling both check groups.

This commit adds the missing comma to properly enable `bugprone-*` checks,
and fixes all resulting warnings across 78 files. The `misc-*` checks are
intentionally removed for now (to be addressed separately).

Changes:
- Fix missing comma after bugprone-* in .clang-tidy
- Remove misc-* from enabled checks (deferred to separate PR)
- Disable bugprone-easily-swappable-parameters (not fixable without API changes)
- Disable bugprone-unchecked-optional-access (false positives with ASSERT_TRUE guards)
- Fix bugprone-macro-parentheses: wrap macro args in parens where safe,
  use NOLINTBEGIN/END for stream-expression macros (Logger.hpp, PluginHelpers.hpp)
- Fix bugprone-narrowing-conversions with explicit static_cast
- Fix bugprone-multi-level-implicit-pointer-conversion with static_cast<void*>
- Fix bugprone-branch-clone by combining duplicate switch cases
- Fix bugprone-fold-init-type with correct accumulator types
- Fix bugprone-string-constructor and bugprone-stringview-nullptr issues
- Fix readability-redundant-access-specifiers in Attributes.hpp and Node.hpp

+++ projects/hipdnn/data_sdk/include/hipdnn_data_sdk/types/Fp8E4M3.hpp
@@ -114,7 +114,8 @@ inline uint8_t float_to_fp8_e4m3_bits(float f, bool saturate = true) noexcept
     std::memcpy(&bits, &f, sizeof(float));
 
     uint32_t sign = (bits >> 24) & 0x80; // Extract sign to bit 7
-    int32_t exp = ((bits >> 23) & 0xFF) - 127 + 7; // Rebias from float (127) to E4M3 (7)
+    int32_t exp = static_cast<int32_t>(((bits >> 23) & 0xFF) - 127
+                                       + 7); // Rebias from float (127) to E4M3 (7)
     uint32_t mant = bits & 0x007FFFFF;
 
     // Handle overflow
